pyproject.toml
[tool.poetry]
name = "ml-fastapi"
version = "0.1.0"
description = "ML project with FastAPI interface"
authors = ["Alexander <as.frantsev@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "ml_fastapi"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
joblib = "^1.3.2"
python-dotenv = "^1.0.0"
jinja2 = "^3.1.2"
loguru = "^0.7.2"
pandas = "^2.1.3"
scikit-learn = "^1.3.2"

[tool.poetry.group.dev.dependencies]
pandas = "^2.1.3"
scikit-learn = "^1.3.2"
ipykernel = "^6.27.1"
httpx = "^0.25.2"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
